+++ src/main/java/com/example/javagyaknb1/repository/CrudLabdarugoRepository.java
@@ -3,5 +3,5 @@ package com.example.javagyaknb1.repository;
 import com.example.javagyaknb1.model.CrudLabdarugo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CrudLabdarugoRepository extends JpaRepository<CrudLabdarugo, Integer> {
+public interface CrudLabdarugoRepository extends JpaRepository<CrudLabdarugo, Long> {
 }
